# Add option to limit thread usage per query