--- a/pinot-core/src/main/java/org/apache/pinot/core/operator/combine/DistinctCombineOperator.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/operator/combine/DistinctCombineOperator.java
@@ -37,8 +37,8 @@ public class DistinctCombineOperator extends BaseCombineOperator {
   private final boolean _hasOrderBy;
 
   public DistinctCombineOperator(List<Operator> operators, QueryContext queryContext, ExecutorService executorService,
-      long endTimeMs) {
-    super(operators, queryContext, executorService, endTimeMs);
+      long endTimeMs, int maxExecutionThreads) {
+    super(operators, queryContext, executorService, endTimeMs, maxExecutionThreads);
     _hasOrderBy = queryContext.getOrderByExpressions() != null;
   }