… pinot-core/src/main/java/org/apache/pinot/core/plan/maker/InstancePlanMakerImplV2.java
… pinot-core/src/test/java/org/apache/pinot/core/operator/combine/CombineSlowOperatorsTest.java
… pinot-core/src/test/java/org/apache/pinot/core/operator/combine/SelectionCombineOperatorTest.java
… pinot-core/src/test/java/org/apache/pinot/core/plan/CombinePlanNodeTest.java
… pinot-core/src/test/java/org/apache/pinot/core/query/aggregation/groupby/GroupByTrimTest.java
… pinot-spi/src/main/java/org/apache/pinot/spi/utils/CommonConstants.java