@xiangfu0 2021-09-29
This is awesome!

One high level question, shall we consider set this numExecutionThreads part of QueryContext?