@mqliang 2021-09-29
@snleee @Jackie-Jiang 
> I think that this is an interesting feature for the multi-tenant cluster.

Do we have admission control mechanism? In multi-tenant environment, if we don't have good admission controlling, a greedy user will set the value of query level option "maxExecutionThreads" as an extremely large value to get the best performance (even if they don't need that), and impact other users. What made matters worse was that, once other users aware that heir performance drops due to a greedy user “stealing” the public resource, they will also increase their config value. Which will eventually burn out the server.