@snleee 2021-09-29
@mqliang Our old behavior is to use `min(numOperator, MAX_NUM_THREADS_PER_QUERY)` for all incoming queries. So, even if the user tries to be greedy (I hope this is not the case in reality), they will get the same behavior as before. I think that we can consider enabling the `server` level config by default for our large MT cluster given that the latency requirement is not too strict. This can protect the cluster from a single super expensive query. In our current scheme, one super-expensive query can bring down the cluster. I think that it's already a good start.

Yes, if we want to solve the problem of "fairness" among all clients for our MT cluster resource, a more sophisticated scheduling algorithm will be needed and this would be a complex research problem.