@Jackie-Jiang 2021-09-29
@mqliang Good point. I considered that and we want to add admission control so that regular users cannot set the query options. This PR won't make things worse because the group-by queries (most expensive ones) are already using all the system resources.
We should build all the knobs so that admin can control how the query should be executed, but also build mechanism to limit regular users from overriding these configs.