@mqliang 2021-09-29
@snleee 
>  if we want to solve the problem of fairness among all clients for our MT cluster resource, a more sophisticated scheduling algorithm will be needed and this would be a complex research problem.

A proactive scheduling algo to ensure "fairness" in the first place may be a overkill, but a protective admission control is helpful to protect the whole cluster and each server.

@Jackie-Jiang 
>  but also build mechanism to limit regular users from overriding these configs.

This is a Authorization(AuthZ), which is a different story from admission control.