@snleee 2021-09-30
@mqliang I think that this is a good topic for discussion. Let's create an issue to capture this. Can you elaborate more about `admission control vs authorization` in the issue? Since this PR is already closed, it will be a bit hard to continue the discussion here.