@mcvsubbu 2021-09-30
Just going by the comments here, it may be good to impose that the per-query value of maxthreads is not above the instance config (i.e. the instance config is the max that anyone can ever set). I am not sure if this check is already being performed. @Jackie-Jiang